Fix float32 misread of CurveData.data_binary in zonation/correlation
CurveData.data_binary is written as float64 bytes everywhere (see
main.py's arr.tobytes() call sites). zonation.py's net-pay curve
extraction and correlation.py's cross-section/correlation-panel curve
loader were both reading it back with dtype=np.float32, which
misinterprets the byte buffer: it silently produces an array of twice
the true length full of garbage reinterpreted values instead of an
error, corrupting net-pay calculations and well-to-well correlation.

Same class of bug as the float32/float64 dtype fix already applied to
qc.py earlier this session. Other float32 usages left alone (verified
against how the data was written): those are locally-computed derived
arrays (ratios, linspace depth grids, interpolation output), not reads
of stored curve bytes.

+++ backend/routers/correlation.py
@@ -77,7 +77,7 @@ def _load_curve(db: Session, well_id: int, curve: str) -> Tuple[Optional[LogRun]
     if not cd or not cd.data_binary:
         return lr, depth, None, None
 
-    arr = np.frombuffer(cd.data_binary, dtype=np.float32)
+    arr = np.frombuffer(cd.data_binary, dtype=np.float64)
     n = min(len(arr), len(depth))
     if n <= 1:
         return lr, depth, None, cd.unit or ""
